backend/requirements.txt
# requirements.txt
fastapi
uvicorn
python-multipart
requests
openai
faiss-cpu
PyPDF2
python-dotenv
numpy
